refactored blocks into separate files
