+++ network-api/networkapi/wagtailpages/customblocks/air_table_block.py
@@ -1,0 +1,15 @@
+from wagtail.core import blocks
+
+
+class AirTableBlock(blocks.StructBlock):
+    url = blocks.URLBlock(
+        help_text="Copied from the Airtable embed code. The word 'embed' will be in the url"
+    )
+    height = blocks.IntegerBlock(
+        default=533,
+        help_text="The height of the view on a desktop, usually copied from the Airtable embed code",
+    )
+
+    class Meta:
+        icon = 'placeholder'
+        template = 'wagtailpages/blocks/airtable_block.html'
