+++ network-api/networkapi/wagtailpages/customblocks/annotated_image_block.py
@@ -1,3 +1,19 @@
+from wagtail.images.blocks import ImageChooserBlock
+from wagtail.core import blocks
+
+
+class ImageBlock(blocks.StructBlock):
+    image = ImageChooserBlock()
+    altText = blocks.CharBlock(
+        required=True,
+        help_text='Image description (for screen readers).'
+    )
+
+    class Meta:
+        icon = 'image'
+        template = 'wagtailpages/blocks/image_block.html'
+
+
 class AnnotatedImageBlock(ImageBlock):
     caption = blocks.CharBlock(
         required=False
