+++ network-api/networkapi/wagtailpages/customblocks/iframe_block.py
@@ -1,0 +1,17 @@
+from wagtail.core import blocks
+
+
+class iFrameBlock(blocks.StructBlock):
+    url = blocks.CharBlock(
+        help_text='Please note that only URLs from white-listed domains will work.'
+    )
+    caption = blocks.CharBlock(
+        required=False
+    )
+    captionURL = blocks.CharBlock(
+        required=False,
+        help_text='Optional URL that this caption should link out to.'
+    )
+
+    class Meta:
+        template = 'wagtailpages/blocks/iframe_block.html'
